Change np.float => np.float32 since it's deprecated in the numpy>1.20.0

+++ pythreshold/global_th/otsu.py
@@ -30,7 +30,7 @@ def otsu_threshold(image=None, hist=None):
 
     # Calculating histogram
     if not hist:
-        hist = np.histogram(image, bins=range(256))[0].astype(np.float)
+        hist = np.histogram(image, bins=range(256))[0].astype(np.float32)
 
     cdf_backg = np.cumsum(np.arange(len(hist)) * hist)
     w_backg = np.cumsum(hist)  # The number of background pixels
@@ -137,7 +137,7 @@ def otsu_multithreshold(image=None, hist=None, nthrs=2):
 
     # Calculating histogram
     if not hist:
-        hist = np.histogram(image, bins=range(256))[0].astype(np.float)
+        hist = np.histogram(image, bins=range(256))[0].astype(np.float32)
 
     # Cumulative histograms
     c_hist = np.cumsum(hist)
