+++ pythreshold/local_th/bradley_roth.py
@@ -32,9 +32,9 @@ def bradley_roth_threshold(img, w_size=15, w=0.15):
 
     # Computing integral image
     # Leaving first row and column in zero for convenience
-    integ = np.zeros((i_rows, i_cols), np.float)
+    integ = np.zeros((i_rows, i_cols), np.float32)
 
-    integ[1:, 1:] = np.cumsum(np.cumsum(img.astype(np.float), axis=0), axis=1)
+    integ[1:, 1:] = np.cumsum(np.cumsum(img.astype(np.float32), axis=0), axis=1)
 
     # Defining grid
     x, y = np.meshgrid(np.arange(1, i_cols), np.arange(1, i_rows))
